OilPvtMultiplexer: move implementation to impl file
